Return named classes
Better for var_dump and stack traces.
